Updated UI for headref matches that werent completed
