--- apps/frontend/src/app/[locale]/lems/(volunteer)/(dashboard)/head-referee/components/desktop-schedule/index.ts
@@ -1,2 +1,0 @@
-export { DesktopScheduleTable } from './desktop-schedule-table';
-export { MatchRow } from './match-row';
